tcc - ok

+++ app/Http/Controllers/ProductsController.php
@@ -4,22 +4,27 @@ namespace App\Http\Controllers;
 
 use App\Models\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ProductsController extends Controller
 {
     public function index()
     {
+        Gate::authorize('viewAny', Product::class);
         $products = Product::all();
         return view('product.index', compact('products'));
     }
 
     public function create()
     {
+        Gate::authorize('create', Product::class);
         return view('product.create');
     }
 
     public function store(Request $request)
     {
+        Gate::authorize('create', Product::class);
+
         $request->validate([
             'nome'      => 'required|string|max:255',
             'descricao' => 'required|string',
@@ -43,16 +48,20 @@ class ProductsController extends Controller
 
     public function show(Product $product)
     {
+        Gate::authorize('view', $product);
         return view('product.show', compact('product'));
     }
 
     public function edit(Product $product)
     {
+        Gate::authorize('update', $product);
         return view('product.edit', compact('product'));
     }
 
     public function update(Request $request, Product $product)
     {
+        Gate::authorize('update', $product); // corrigido
+
         $request->validate([
             'nome'      => 'required|string|max:255',
             'descricao' => 'required|string',
@@ -76,6 +85,7 @@ class ProductsController extends Controller
 
     public function destroy(Product $product)
     {
+        Gate::authorize('delete', $product); // primeiro autorizo
         $product->delete();
 
         return redirect()->route('product.index')->with('success', 'Produto deletado com sucesso!');
